local_app/requirements.txt
pillow~=11.0.0
Flask~=3.1.0
PyAudioWPatch~=0.2.12.6
vosk~=0.3.45